requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.0
Flask-Login>=0.5
Flask-WTF>=1.0
Flask-Bcrypt>=1.0
PyMySQL>=1.0
python-dotenv>=0.19
gunicorn>=20.0 # Para deployment
email-validator>=1.1 # Dependência do Flask-WTF para validação de email